Fix compilation errors due to refactoring.
